mirror playwright's screenshot options (#1233)
# why
- to align stagehand's `page.screenshot()` function to with the familiar
`page.screenshot()` function from playwright
# what changed
- added support for the following arguments:
- `animations`, can be `"disabled" | "allow"`
   - if disabled, `disableAnimations()` is called
- `disableAnimations()` injects CSS styles to cancel or finish
animations that are in progress. it returns a callback that cleans up
the injected CSS when it is invoked.
- `caret`, can be `"hide" | "initial"`
   - if `"hide"` is passed in, then `hideCaret()` is called
- `hideCaret()` injects CSS styles which makes the caret appear as
transparent. it returns a callback that cleans up the injected CSS when
it is invoked
- `clip`, an object with `x`, `y`, `width` & `height` args
- if `clip` is provided, it is passed to the underlying CDP
`Page.captureScreenshot` call. this will screenshot only the region
specified by the `clip`
- `fullPage` is directly mapped to the `captureBeyondViewport` arg in
the CDP `Page.captureScreenshot` call
- `mask`, an array of `locator`'s which point to nodes that will be
covered with an overlay
- `maskColor`, a CSS colour for the masked nodes
- `omitBackground`, whether the background for translucent nodes should
be white, or just omitted
- `path`, the filepath to where the screenshot should be saved
- `quality`, 0-100 of how high quality the screenshot should be. only
works for jpeg
- `scale`, can be `"css" | "device"`. whether the screenshot should be
rendered as 1 pixel per css pixel, or if the device pixel ratio should
be used
- `style`, additional CSS styles that should be injected before the
screenshot is taken. removed after the screenshot has finished.
- `timeout`, timeout in MS
- `type`, whether the screenshot should be `"jpeg"` or `"png"`
# test plan
- added unit testa