Add minimal contract test for public api (#1255)
# why
We want to be able to see what the public API shape looks like - which
top level functions are exported.

# what changed
Added vitest and a minimal test file that checks which functions are
exported at the root level. If new functions are exported or removed,
this test will catch that. I also added one example of drilling down and
testing the parameters of StagehandMetrics — in the future we might want
to do this for all exports.

Curious if this surfaces and exports that don't need to be exported or
not.

# test plan
Just added tests. They run in the CI pipeline or can be run with `pnpm
--filter @browserbasehq/stagehand run test:vitest`