add BB session URL & debug URL accessors (#1281)
# why
currently, stagehand just logs the BB session URL and debug URL inline
without storing this info on the stagehand object but we want getters
for these.

If env = browserbase, we return the urls, if it's local we return an
empty string currently as per the ticket. It seemed to me like returning
undefined if env = local might be better DX, but i'll double check w/
sean.

# what changed
Added getters and a small `resetBrowserbaseSessionMetadata()` helper.

# test plan
Adde vitest tests that show that this works when running with `env:
"BROWSERBASE"` and returns an empty string when `env: "LOCAL"`.