Add public API contract tests (#1290)
# why
We need to create stubs for the client sdk, and this allows us to write
tests so that we can confirm that our stub signatures have full parity.

# what changed
Added full contract tests which check just the public facing API *types*
https://vitest.dev/guide/testing-types

Also made it so that all types from `/public/types` are explicitly
re-exported through index.ts, so that we can have a single place to
review all public exports.

These tests might be a bit more in-depth then we ultimately want - we
are checking function signatures for stuff like the vercel AI client,
which might be more brittle than we want currently, but we will likely
be removing some of these public exports anyways with the canonical
work.

# test plan
You can run `pnpm --filter @browserbasehq/stagehand test:vitest`